Modified for github pages hosting
