--- qfin-website/src/pages/about.js
@@ -1,17 +1,0 @@
-import React from 'react';
-  
-const About = () => {
-  return (
-    <div className="page"
-    style={{
-      height: '100vh',
-    }}     
-    >
-      <h1>Info about QFin</h1>
-      <p>QFin be a club tho</p>
-      <p>Qfin2</p>
-    </div>
-  );
-};
-  
-export default About;
