--- qfin-website/src/pages/blogs.js
@@ -1,19 +1,0 @@
-
-import React from 'react';
-  
-const Blogs = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>Welcome to QFin Blogs</h1>
-    </div>
-  );
-};
-  
-export default Blogs;
