--- qfin-website/src/pages/events.js
@@ -1,19 +1,0 @@
-
-import React from 'react';
-  
-const Events = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>QFin past and future Events</h1>
-    </div>
-  );
-};
-  
-export default Events;
