--- qfin-website/src/pages/index.js
@@ -1,18 +1,0 @@
-import React from 'react';
-  
-const Home = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>Welcome to QFin</h1>
-    </div>
-  );
-};
-  
-export default Home;
