--- qfin-website/src/pages/projects.js
@@ -1,19 +1,0 @@
-
-import React from 'react';
-  
-const Projects = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>QFin Projects</h1>
-    </div>
-  );
-};
-  
-export default Projects;
