--- qfin-website/src/pages/signup.js
@@ -1,19 +1,0 @@
-
-import React from 'react';
-  
-const SignUp = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>QFin Sign Up</h1>
-    </div>
-  );
-};
-  
-export default SignUp;
