--- qfin-website/src/pages/team.js
@@ -1,19 +1,0 @@
-
-import React from 'react';
-  
-const Teams = () => {
-  return (
-    <div className="page"
-      style={{
-        display: 'flex',
-        justifyContent: 'Left',
-        alignItems: 'Left',
-        height: '100vh'
-      }}
-    >
-      <h1>Welcome to QFin Team</h1>
-    </div>
-  );
-};
-  
-export default Teams;
