+++ src/pages/blogs.js
@@ -1,0 +1,15 @@
+
+import React from 'react';
+  
+const Blogs = () => {
+  return (
+    <div className="page">
+      <h1 className="title">Welcome to QFin Blogs</h1>
+      <p className="content">QFin be a club tho</p>
+      <p className="content">"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."</p>
+
+    </div>
+  );
+};
+  
+export default Blogs;
