+++ src/pages/events.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+  
+const Events = () => {
+  return (
+    <div className="page">
+      <h1 className="title">QFin past and future Events</h1>
+    </div>
+  );
+};
+  
+export default Events;
