+++ src/pages/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+  
+const Home = () => {
+  return (
+    <div className="page">
+      <h1 className="title">Welcome to QFin</h1>
+    </div>
+  );
+};
+  
+export default Home;
