+++ src/pages/projects.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+  
+const Projects = () => {
+  return (
+    <div className="page">
+      <h1 className="title">QFin Projects</h1>
+    </div>
+  );
+};
+  
+export default Projects;
