+++ src/pages/signup.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+  
+const SignUp = () => {
+  return (
+    <div className="page">
+      <h1 className="title">QFin Sign Up</h1>
+    </div>
+  );
+};
+  
+export default SignUp;
