+++ src/pages/team.js
@@ -1,0 +1,12 @@
+
+import React from 'react';
+  
+const Teams = () => {
+  return (
+    <div className="page">
+      <h1 className="title">Welcome to QFin Team</h1>
+    </div>
+  );
+};
+  
+export default Teams;
